requirements.txt
cryptography
pytest
hvac
keyring
cryptography
pyyaml